Originator's sample didn't work on Mac. Removed Makefile and replaced with shell script

+++ Sridhar.c
@@ -10,7 +10,7 @@ JNIEXPORT void JNICALL Java_Sridhar_printx (JNIEnv *env, jobject obj) {
 	// Call an existing library (from time.h)
 	struct timeval te; 
 	gettimeofday(&te, NULL); // get current time
-	long milliseconds = te.tv_sec*1000LL + te.tv_usec/1000; // caculate milliseconds
+	long long milliseconds = te.tv_sec*1000LL + te.tv_usec/1000; // caculate milliseconds
 	printf(".c :: Java_Sridhar_print() :: current time is %lld milliseconds\n", milliseconds);
 }
